+++ components/tree/check-props.js
@@ -8,52 +8,21 @@ Neither the name of salesforce.com, inc. nor the names of its contributors may b
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
+/* eslint-disable import/no-mutable-exports */
 
-// ### React
-// React is an external dependency of the project.
-import React from 'react';
+import oneOfRequiredProperty from '../../utilities/warning/one-of-required-property';
 
-import { CARD_BODY } from '../../utilities/constants';
+let checkProps = function () {};
 
-// ### classNames
-// [github.com/JedWatson/classnames](https://github.com/JedWatson/classnames)
-// This project uses `classnames`, "a simple javascript utility for conditionally joining classNames together."
-import classNames from 'classnames';
+if (process.env.NODE_ENV !== 'production') {
+	checkProps = function (COMPONENT, props) {
+		/* eslint-disable max-len */
+		oneOfRequiredProperty(COMPONENT, {
+			assistiveText: props.assistiveText,
+			heading: props.heading
+		});
+		/* eslint-enable max-len */
+	};
+}
 
-const cssClasses = {
-	base: 'slds-card__body'
-};
-
-const idSuffixes = {
-	base: '__body'
-};
-
-const CardBody = (props) => (
-	<div
-		className={classNames(cssClasses.base, props.className)}
-		id={props.id + idSuffixes.base}
-	>
-		{props.children}
-	</div>
-);
-
-CardBody.displayName = CARD_BODY;
-
-CardBody.propTypes = {
-	/**
-	 * Elements to place in the body.
-	 */
-	children: React.PropTypes.node,
-	/**
-	 * CSS classes to be added to the card.
-	 */
-	className: React.PropTypes.oneOfType([React.PropTypes.array, React.PropTypes.object, React.PropTypes.string]),
-	/**
-	 * Set the HTML `id` of the body.
-	 */
-	id: React.PropTypes.string
-};
-
-module.exports = CardBody;
-module.exports.cssClasses = cssClasses;
-module.exports.idSuffixes = idSuffixes;
+export default checkProps;