+++ stories/app-launcher/index.jsx
@@ -41,7 +41,8 @@ const DemoAppLauncherTile = React.createClass({
 				title="Marketing Cloud"
 				iconText="MC"
 				description="Send emails, track emails, read emails! Emails!"
-				onClick={action('Tile clicked!')}
+				href="https://www.marketingcloud.com/"
+				onClick={action('Tile clicked! Actual href should be ignored')}
 				search={this.props.search}
 				size={this.props.size}
 			/>
@@ -79,6 +80,7 @@ const DemoAppLauncherTileWithIconNode = React.createClass({
 			<AppLauncherTile
 				title="Sales Cloud"
 				description="The primary internal Salesforce org."
+				href="https://www.salesforce.com/"
 				iconNode={icon}
 				onClick={action('Tile with icon node clicked!')}
 				search={this.props.search}
@@ -250,7 +252,7 @@ const DemoAppLauncher = React.createClass({
 
 	render () {
 		const search = <Search onChange={this.onSearch} placeholder="Find an app" assistiveText="Find an app" />;
-		const modalHeaderButton = <Button label="App Exchange" onclick={action('Modal Button clicked!')} />;
+		const modalHeaderButton = <Button label="App Exchange" onClick={action('Modal Button clicked!')} />;
 
 		return (
 			<GlobalNavigationBar>